Small fixes/improvements

+++ Eos/src/Platform/Windows/WindowsPlatformUtils.cpp
@@ -56,7 +56,7 @@ namespace Eos {
 
 	float Time::GetTime()
 	{
-		return glfwGetTime();
+		return (float)glfwGetTime();
 	}
 
 }
